+++ PayPal/src/main/java/com/braintreepayments/api/PayPalFlowStartedCallback.java
@@ -5,14 +5,13 @@ import androidx.fragment.app.FragmentActivity;
 
 /**
  * Callback for receiving result of
- * {@link PayPalClient#requestOneTimePayment(FragmentActivity, PayPalCheckoutRequest, PayPalFlowStartedCallback)} and
- * {@link PayPalClient#requestBillingAgreement(FragmentActivity, PayPalVaultRequest, PayPalFlowStartedCallback)} and
  * {@link PayPalClient#tokenizePayPalAccount(FragmentActivity, PayPalRequest, PayPalFlowStartedCallback)}.
  */
 public interface PayPalFlowStartedCallback {
 
     /**
-     * @param error an exception that occurred while initiating a PayPal transaction
+     * @param payPalResponse the result of the PayPal web authentication flow
+     * @param error          an exception that occurred while initiating a PayPal transaction
      */
-    void onResult(@Nullable Exception error);
+    void onResult(PayPalResponse payPalResponse, @Nullable Exception error);
 }